safari-600.7-branch/LayoutTests/platform/ios-sim/http/tests/cache/xhr-vary-header-expected.txt

Test that XMLHttpRequest doesn't return a wrong cached answer when there is a Vary header in response.

Make cross origin request 
 PASS responseText is 'Cross origin response'
PASS responseText is 'Same origin response'
PASS responseText is 'Cross origin response'
PASS successfullyParsed is true

TEST COMPLETE
